Conflicts:
	mysql-test/include/commit.inc
	mysql-test/r/commit_1innodb.result
	mysql-test/r/create.result
	mysql-test/r/create_ps_protocol.result
	mysql-test/r/group_min_max_ps_protocol.result
	mysql-test/r/partition_explicit_prune.result
	mysql-test/r/partition_locking.result
	mysql-test/r/partition_locking_ps_protocol.result
	mysql-test/suite/perfschema/r/alter_table_progress.result
	sql/dd/dd_schema.cc
	sql/dd/dd_table.cc
	sql/sql_table.cc

+++ sql/item_func.h
@@ -1931,7 +1931,10 @@ public:
     :Item_udf_func(pos, udf_arg, opt_list)
   {}
   longlong val_int() override;
-  double val_real() override { return Item_func_udf_int::val_int(); }
+  double val_real() override
+  {
+    return static_cast<double>(Item_func_udf_int::val_int());
+  }
   String *val_str(String *str) override;
   bool get_date(MYSQL_TIME *ltime, my_time_flags_t fuzzydate) override
   {
